==> codigo_principal/2_OCR.ipynb <==
#### IMPORTAR

import matplotlib.pyplot as plt
from doctr.io import DocumentFile
from doctr.models import ocr_predictor
import os
import glob

data_dir = r"RUTA"

pdf_files = glob.glob(os.path.join(data_dir, "*.pdf"))
print(f"Found {len(pdf_files)} PDF files.")

txt_files = glob.glob(os.path.join(data_dir, "*.txt"))
print(f"Found {len(txt_files)} TXT files.")

### PREDICTOR

predictor = ocr_predictor(pretrained=True)
print(predictor)

os.makedirs(data_dir, exist_ok=True)

for pdf_file in pdf_files[:]:
  print(pdf_file)

for txt_file in txt_files[:]:
  print(txt_file)

os.makedirs(data_dir, exist_ok=True)
default_page_ranges = [
    (0, 50, "parte1"),
    (100, 150, "parte2"),
    (150, 200, "parte3"),
    (200, 250, "parte4"),
    (250, 300, "parte5"),
    (300, 350, "parte6"),
    (350, 400, "parte7"),
    (400, 450, "parte8"),
    (450, 500, "parte9"),
    (500, 550, "parte10"),
    (550, 600, "parte11"),
    (600, 650, "parte12"),
    (650, 700, "parte13"),
    (700, 750, "parte14"),
    (750, 800, "parte15"),
    (800, 850, "parte16"),
    (850, 900, "parte17"),
]

for pdf_file in pdf_files[:]:  
    print(f"Processing file: {pdf_file}")

    name = os.path.splitext(os.path.basename(pdf_file))[0]
    
    print("Escaneado")
    if pdf_file == "RUTA":
        page_ranges = default_page_ranges[1:] 
    else:
        page_ranges = default_page_ranges

    try:
        doc = DocumentFile.from_pdf(pdf_file)
        print(f"Number of pages: {len(doc)}")

        for start, end, suffix in page_ranges:
            if start >= len(doc):
                print(f"Skipping range {start}-{end} as document only has {len(doc)} pages")
                continue

            print(f"Processing pages {start + 1} to {min(end, len(doc))}")
            page_texts = []

            for page_num in range(start, min(end, len(doc))):
                print(f"Processing page {page_num + 1}/{len(doc)}")
                page = doc[page_num]

                result = predictor([page])
                page_text = result.render()
                page_texts.append(page_text)

            if page_texts:
                text_output = '\n'.join(page_texts)
                base_name = os.path.splitext(os.path.basename(pdf_file))[0]
                output_filename = f"{base_name}_TXT.txt"
                output_filepath = os.path.join(data_dir, output_filename)

                with open(output_filepath, 'w', encoding='utf-8') as f:
                    f.write(text_output)

                print(f"Text output saved to {output_filepath}\n")

    except Exception as e:
        print(f"An error occurred while processing {pdf_file}: {e}\n")
